Support ISO-3166-1 Alpha-3 country codes
This commit adds support for the full 3 letter alpha-3 codes as defined
by the ISO-3166-1 standard.

As per our previous commits we deviate from the standard in including
the country data for Kosovo.

This commit uses XKO as the alpha-3 code for Kosovo. This was chosen by
triaging data from multiple sources online - this seems to be the most
popular and widely used code representing Kosovo at this point.

The other reasonably popular option is XKX, But I found more instances
of people using XKO

¯\(º_o)/¯

--

1.
http://www.statcan.gc.ca/eng/subjects/standard/sccai/2011/scountry-desc
2. http://www.agrodep.org/country/xko